+++ examples/matrices.js
@@ -82,14 +82,3 @@ print(math.range(0, 18, 3));                // [0, 3, 6, 9, 12, 15]
 print(math.range('2:-1:-3'));               // [2, 1, 0, -1, -2]
 print(math.factorial(math.range('1:6')));   // [1, 2, 6, 24, 120]
 console.log();
-
-console.log('use the expression parser');
-var parser = math.parser();
-print(parser.eval('d = [1, 2; 3, 4]'));     // [[1, 2], [3, 4]]
-print(parser.eval('e = zeros(2, 2)'));      // [[0, 0], [0, 0]]
-print(parser.eval('e[0, 0:2] = [5, 6]'));   // [[5, 6], [0, 0]]
-print(parser.eval('e[1, :] = [7, 8]'));     // [[5, 6], [7, 8]]
-print(parser.eval('f = d * e'));            // [[19, 22], [43, 50]]
-print(parser.eval('g = f[1, 0]'));          // 43
-print(parser.eval('g = f[:, 0]'));          // [[19], [43]]
-console.log();