+++ test/function/construction/string.test.js
@@ -7,7 +7,7 @@ describe('string', function() {
   it('should be parsed correctly', function() {
     assert.equal(math.eval('string(123)'), '123');
     assert.equal(math.eval('string(2+3i)'), '2 + 3i');
-    assert.equal(math.eval('string(1:6)'), '[1, 2, 3, 4, 5]');
+    assert.equal(math.eval('string(1:5)'), '[1, 2, 3, 4, 5]');
     assert.equal(math.eval('string(2 inch)'), '2 inch');
     assert.equal(math.eval('string([1,2;3,4])'), '[[1, 2], [3, 4]]');
   });